requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-text-splitters==0.2.2
langchain-groq==0.1.6
transformers==4.43.2
google-generativeai
unstructured==0.15.0
pdfminer.six==20221105
Flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.1
flask-cors 
nltk
faiss-cpu